oppianproj/requirements.txt
BeautifulSoup
Markdown==2.0.3
Pygments==1.3.1
South==0.7
boto==1.9b
python-dateutil==1.5
simplejson==2.1.1
twitter==1.3.1

# this doesn't seem to work on Snow Leopard,
# so moving to the end, so that the other dependencies are installed
django-defaultsite==1.1

